full implementation of document occupation operations

+++ Controllers/DocumentController.cs
@@ -14,7 +14,7 @@ public class DocumentController(IDocumentService _documentService, IOccupancySer
 
     [HttpGet]
     [Authorize]
-    [ValidateUserId]
+    [GetUserId]
     public async Task<IActionResult> GetDocumentTitles([FromQuery] string? query, [FromQuery] int offset, [FromQuery] int quantity)
     {
         if (MaxRequestableDocumentCount < quantity)
@@ -31,7 +31,7 @@ public class DocumentController(IDocumentService _documentService, IOccupancySer
 
     [HttpPost]
     [Authorize]
-    [ValidateUserId]
+    [GetUserId]
     public async Task<IActionResult> AddNewDocument([FromBody] CreateDocumentDto document)
     {
         int userId = (int)HttpContext.Items["UserId"]!;
@@ -43,7 +43,8 @@ public class DocumentController(IDocumentService _documentService, IOccupancySer
 
     [HttpPost("{id}")]
     [Authorize]
-    [ValidateUserId]
+    [GetUserId]
+    [ServiceFilter(typeof(ValidateSessionIdAttribute))]
     public async Task<IActionResult> ModifyDocumentText(int id, [FromBody] ModifyDocumentDto document)
     {
         int userId = (int)HttpContext.Items["UserId"]!;
@@ -58,7 +59,8 @@ public class DocumentController(IDocumentService _documentService, IOccupancySer
 
     [HttpGet("{id}")]
     [Authorize]
-    [ValidateUserId]
+    [GetUserId]
+    [ServiceFilter(typeof(ValidateSessionIdAttribute))]
     public async Task<IActionResult> GetDocumentById(int id)
     {
         int userId = (int)HttpContext.Items["UserId"]!;
@@ -68,12 +70,57 @@ public class DocumentController(IDocumentService _documentService, IOccupancySer
         return result is not null ? Ok(result) : NotFound();
     }
 
+    [HttpPost("{id}/occupy")]
+    [Authorize]
+    [GetUserId]
+    [GetSessionId]
+    public async Task<IActionResult> OccupyDocument(int id)
+    {
+        int userId = (int)HttpContext.Items["UserId"]!;
+        long sessionId = (long)HttpContext.Items["SessionId"]!;
+
+        var tryOccupyDocumentResult = _occupancyService.TryOccupyDocument(userId, sessionId, id);
+
+        return tryOccupyDocumentResult switch
+        {
+            DocumentLockOpResult.INVALID_SESSION_ID or DocumentLockOpResult.SESSION_ID_OWNER_DOES_NOT_MATCH => Unauthorized(),
+            DocumentLockOpResult.DOCUMENT_OCCUPIED => Conflict(),
+            DocumentLockOpResult.SUCCESS => Ok(),
+            _ => throw new Exception($"Unexpected result {tryOccupyDocumentResult} returned from TryOccupyDocument()."),
+        };
+    }
+
+    [HttpPost("{id}/unoccupy")]
+    [Authorize]
+    [GetUserId]
+    [ServiceFilter(typeof(ValidateSessionIdAttribute))]
+    public async Task<IActionResult> UnoccupyDocument(int id)
+    {
+        int userId = (int)HttpContext.Items["UserId"]!;
+        long sessionId = (long)HttpContext.Items["SessionId"]!;
+
+        var tryUnoccupyDocumentResult = _occupancyService.TryRemoveDocumentLock(userId, sessionId, id);
+
+        return tryUnoccupyDocumentResult switch
+        {
+            DocumentLockOpResult.DOCUMENT_UNOCCUPIED => BadRequest(),
+            DocumentLockOpResult.SUCCESS => Ok(),
+            _ => throw new Exception($"Unexpected result {tryUnoccupyDocumentResult} returned from TryRemoveDocumentLock()."),
+        };
+    }
+
     [HttpDelete("{id}")]
     [Authorize]
-    [ValidateUserId]
+    [GetUserId]
+    [ServiceFilter(typeof(ValidateSessionIdAttribute))]
     public async Task<IActionResult> DeleteDocumentById(int id)
     {
         int userId = (int)HttpContext.Items["UserId"]!;
+
+        var isDocumentOccupiedResult = (DocumentLockOpResult)HttpContext.Items["SessionState"]!;
+
+        if (isDocumentOccupiedResult == DocumentLockOpResult.DOCUMENT_OCCUPIED_BY_SESSION)
+            return Conflict();
 
         var result = await _documentService.DeleteDocument(id, userId);
 
@@ -82,7 +129,7 @@ public class DocumentController(IDocumentService _documentService, IOccupancySer
 
     [Route("ws")]
     [Authorize]
-    [ValidateUserId]
+    [GetUserId]
     public async Task DocumentStateWebsocket()
     {
         if (!HttpContext.WebSockets.IsWebSocketRequest)
@@ -100,7 +147,7 @@ public class DocumentController(IDocumentService _documentService, IOccupancySer
 
     [HttpPost("mock")]
     [Authorize]
-    [ValidateUserId]
+    [GetUserId]
     public async Task<IActionResult> CreateMockDocuments([FromQuery] int quantity)
     {
         int userId = (int)HttpContext.Items["UserId"]!;
